🧪 Move myst-to-tex tests to yml
